requirements.txt
Django>=5.1
# APIs
djangorestframework>=3.15

# Async & workers
channels>=4.0
channels-redis>=4.2
celery>=5.4
redis>=5.0

# DB
psycopg[binary]>=3.1

# Utils
python-dotenv>=1.0
pillow>=10.3
uvicorn[standard]>=0.30
